Create identifySource.js
This file creates a cookie that identifies traffic source based on a query string appended to any inbound traffic links.  This can then be used to set a variable within GTM that can be used in any Tag requiring source to be set.  The catch all else statement sets the traffic source to an arbitry value of 'aw' but this can be changed to anything as desired..